+++ src/main/java/com/github/mkram17/bazaarutils/features/gui/buttons/bookmarks/BookmarkSearchWidget.java
@@ -2,6 +2,7 @@ package com.github.mkram17.bazaarutils.features.gui.buttons.bookmarks;
 
 import com.github.mkram17.bazaarutils.config.features.gui.ButtonsConfig;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.OrderUtil;
+import com.github.mkram17.bazaarutils.utils.minecraft.SlotLookup;
 import com.github.mkram17.bazaarutils.utils.minecraft.gui.widgets.ItemSlotButtonWidget;
 import com.github.mkram17.bazaarutils.utils.PlayerActionUtil;
 import com.github.mkram17.bazaarutils.utils.SoundUtil;
@@ -11,8 +12,6 @@ import com.github.mkram17.bazaarutils.utils.bazaar.gui.BazaarScreenType;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.order.TransactionType;
 import com.github.mkram17.bazaarutils.utils.bazaar.market.price.PricingPosition;
 import com.github.mkram17.bazaarutils.utils.minecraft.item.ItemButton;
-import com.github.mkram17.bazaarutils.utils.minecraft.PlayerSlots;
-import com.github.mkram17.bazaarutils.utils.minecraft.gui.ScreenType;
 import com.github.mkram17.bazaarutils.utils.minecraft.gui.container.ContainerManager;
 import com.github.mkram17.bazaarutils.utils.minecraft.gui.widgets.WidgetManager;
 import net.minecraft.client.Minecraft;
@@ -85,7 +84,7 @@ public class BookmarkSearchWidget {
     public static void onWidgetLeftClick(Bookmark bookmark) {
         SoundUtil.playSound(ItemButton.BUTTON_SOUND, ItemButton.BUTTON_VOLUME);
 
-        Optional<Integer> inventorySlot = PlayerSlots.findScreenSlotByProductId(bookmark.productID());
+        Optional<Integer> inventorySlot = SlotLookup.findScreenSlotByProductId(bookmark.productID());
 
         if (inventorySlot.isPresent()) {
             ContainerManager.clickSlot(inventorySlot.get(), 0);
